+++ cmd/monolith/main.go
@@ -20,14 +20,14 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/google/exposure-notifications-server/internal/api/cleanup"
-	"github.com/google/exposure-notifications-server/internal/api/export"
-	"github.com/google/exposure-notifications-server/internal/api/federationin"
-	"github.com/google/exposure-notifications-server/internal/api/handlers"
-	"github.com/google/exposure-notifications-server/internal/api/publish"
+	"github.com/google/exposure-notifications-server/internal/cleanup"
 	"github.com/google/exposure-notifications-server/internal/database"
 	"github.com/google/exposure-notifications-server/internal/dbapiconfig"
+	"github.com/google/exposure-notifications-server/internal/export"
+	"github.com/google/exposure-notifications-server/internal/federationin"
+	"github.com/google/exposure-notifications-server/internal/handlers"
 	"github.com/google/exposure-notifications-server/internal/logging"
+	"github.com/google/exposure-notifications-server/internal/publish"
 	"github.com/google/exposure-notifications-server/internal/setup"
 )
